requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
braceexpand==0.1.7
chardet==4.0.0
click==8.0.1
discord-custom-help==0.0.6
discord.py==1.7.3
humanize==3.11.0
idna==3.2
import-expression==1.1.4
jishaku==2.2.0
multidict==5.1.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
timetils==2.0.5
typing-extensions==3.10.0.0
yarl==1.6.3
